Creacion inicial de endpoint y algunas modificaciones en modelos

+++ Backend/Telefonica/applications/Cupos/models.py
@@ -7,7 +7,7 @@ class Cupos(models.Model):
         ("POST", "Postpago")
     ]
     numero_movil = models.CharField(max_length=12)
-    saldo = models.IntegerField()
+    saldo = models.DecimalField(decimal_places=2, max_digits=10)
     datos = models.DecimalField(decimal_places=2, max_digits=10)
     plataforma = models.CharField(max_length=4, choices=PLATAFORMA_CHOICES, default="PRE")
     objects = CuposManager()
